pyproject.toml
[project]
name = "multicalcli"
version = "0.1.0"
description = "Multi-account Google Calendar CLI"
readme = "README.md"
authors = [
    { name = "namhyuk kim", email = "happykimnh@icloud.com" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-auth>=2.0",
    "rich>=13.0",
    "python-dateutil>=2.8",
    "parsedatetime>=2.6",
]

[project.scripts]
mcli = "multicalcli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/multicalcli"]
